specialize `trackAction` to `trackClickActions`
